fix: fix errors in accounting service data type map

+++ Ozym.DAL/BusinessLogic/Accounting/AccountBalanceRecord.cs
@@ -12,7 +12,7 @@ namespace Ozym.BusinessLogic.Accounting
 
         public string DisplayName { get; init; }
 
-        public float? Balance { get; init; }
+        public double? Balance { get; init; }
 
         public DateTime? AsOfDate { get; init; }
     }
